+++ Ambassador/src/utopia/ambassador/database/AuthDbExtensions.scala
@@ -1,7 +1,9 @@
 package utopia.ambassador.database
 
+import utopia.ambassador.database.access.many.scope.DbScopes
 import utopia.ambassador.database.access.many.token.DbAuthTokens
 import utopia.citadel.database.access.single.DbUser.DbSingleUser
+import utopia.citadel.database.access.single.organization.DbTask.DbSingleTask
 import utopia.vault.database.Connection
 
 /**
@@ -25,5 +27,13 @@ object AuthDbExtensions
 		  * @return Ids of the scopes that are currently available for this user without a new OAuth process
 		  */
 		def accessibleScopeIds(implicit connection: Connection) = authTokens.scopeIds
+	}
+	
+	implicit class DbAuthTask(val a: DbSingleTask) extends AnyVal
+	{
+		/**
+		  * @return An access point to this task's scopes
+		  */
+		def scopes = DbScopes.forTaskWithId(a.taskId)
 	}
 }