+++ Exodus/Changes.md
@@ -1,9 +1,14 @@
 # Utopia Exodus - List of Changes
 
 ## v2.0.1 (in development)
+### New Features
+- **TasksNode** now provides access to individual **TaskNode**s which support custom extensions
+- Added **ExtendableSessionResourceFactory** class for easier session-based extensions
 ### New Methods
 - **AuthorizedContext**
   - Added `.handlePossibleValuePost(...)` that allows users to recover from cases where no request body was specified
+- **SessionUseCaseImplementation**
+  - Added `.factory(...)` -method to support session-based extendable resource factories
 
 ## v2.0 - 13.7.2021
 This release separates most of the database interaction into the new **Utopia Citadel** module and also makes